+++ src/source/tuner/fcd/proplusV2/FCD2TunerController.java
@@ -64,7 +64,7 @@ public class FCD2TunerController extends FCDTunerController
         }
     }
 
-    public int getCurrentSampleRate()
+    public double getCurrentSampleRate()
     {
         return SAMPLE_RATE;
     }
